// BDI_back/src/shared/enums/role.enum.ts
export enum UserRole {
    ADMIN = "admin",
    SELLER = "seller",
    USER = "user",
    USER_ADMIN = "user_admin",
    USER_READ = "user_read",
    USER_WRITE = "user_write",
    USER_DELETE = "user_delete",
    USER_POST_ADMIN = "user_post_admin",
    USER_POST_READ = "user_post_read",
    USER_POST_WRITE = "user_post_write",
    USER_POST_DELETE = "user_post_delete",
    PROPERTY_ADMIN = "property_admin",
    PROPERTY_READ = "property_read",
    PROPERTY_WRITE = "property_write",
    PROPERTY_DELETE = "property_delete",
    POST_ADMIN = "post_admin",
    POST_READ = "post_read",
    POST_WRITE = "post_write",
    POST_DELETE = "post_delete",
}